rename typechain folder name
