+++ .gitignore
@@ -108,7 +108,7 @@ cache
 artifacts
 
 # typechain
-typechain/
+typechain-types/
 
 coverage*
