+++ utils/index.ts
@@ -1,5 +1,5 @@
 import { ethers } from "hardhat";
-import { ERC20Mock__factory } from "../typechain";
+import { ERC20Mock__factory } from "../typechain-types";
 
 export const getBigNumber = (amount: number, decimals = 18) => {
   return ethers.utils.parseUnits(amount.toString(), decimals);
